implement review: move block to util class
